finish feature tests and some bug fix

+++ app/Infrastructure/Chat/Repositories/DestroyChatRepository.php
@@ -7,8 +7,17 @@ use App\Domain\Chat\Repositories\DestroyChatRepositoryInterface;
 
 class DestroyChatRepository implements DestroyChatRepositoryInterface
 {
-    public function exec(string $id):bool
+    public function exec(string $id):int
     {
+        $chat = Chat::find($id);
+        if(!$chat){
+            return 0;
+        }
+
+        if (!in_array(auth()->id(), $chat->participants)) {
+            return 0;
+        }
+
         return Chat::destroy($id);
     }
 }
